STEMai-frontend/requirements.txt
Flask==3.1.2
gunicorn==23.0.0
python-dotenv==1.2.1
httpx==0.28.1